chore: Rename internal test export for clarity
