+++ .changeset/seven-shrimps-work.md
@@ -1,0 +1,5 @@
+---
+"@sables/core": patch
+---
+
+Rename internal test export
